Ascii Encryption Python
A few of the characters need to be tweaked but that's not hard to fix and you can easily do it yourself. For example i made a mistake and instead of using ? it shows <. Simple enough to fix, just change them accordingly. It works tho. I'm working on a version 2.0 as we speak. Enjoy!